dss_sharp.Error.UseExceptions: Toggle mapping Error.Number to C# exceptions
Previously, we forced the use of exceptions. Now users can easily toggle
the behavior. The default will still be to use exceptions for the safety
of new users, besides the added convenience.

**Usage:** set `dss_sharp.Error.UseExceptions = false` to deactivate the
automatic mapping. Other toggles as `Error.EarlyAbort` and the upcoming
`DSS.CompatFlags` should be enough to more closely match the logic of
older programs using the official COM DLL behavior even for the more
lax usage of OpenDSS.

Closes https://github.com/dss-extensions/dss_sharp/issues/15